salary and transfer completed

+++ src/app/modules/location/location.service.ts
@@ -1,7 +1,12 @@
 import httpStatus from 'http-status';
 import prisma from '../../../shared/prisma';
-import { Location } from '@prisma/client';
+import { Location, Prisma } from '@prisma/client';
 import ApiError from '../../../errors/ApiError';
+import { ILocationFilters } from './location.interface';
+import { IPaginationOptions } from '../../../interfaces/pagination';
+import { IGenericResponse } from '../../../interfaces/common';
+import { paginationHelpers } from '../../../helpers/paginationHelper';
+import { locationSearchableFields } from './location.constant';
 
 // create Location
 const createLocation = async (data: Location): Promise<Location | null> => {
@@ -14,15 +19,65 @@ const createLocation = async (data: Location): Promise<Location | null> => {
   return result;
 };
 
-// get all Locations
-const getLocations = async (): Promise<Location[]> => {
+// get Locations
+const getLocations = async (
+  filters: ILocationFilters,
+  paginationOptions: IPaginationOptions
+): Promise<IGenericResponse<Location[]>> => {
+  const { searchTerm, ...filterData } = filters;
+  const { page, limit, skip, sortBy, sortOrder } =
+    paginationHelpers.calculatePagination(paginationOptions);
+
+  const andConditions = [];
+
+  if (searchTerm) {
+    andConditions.push({
+      OR: locationSearchableFields.map(field => ({
+        [field]: {
+          contains: searchTerm,
+        },
+      })),
+    });
+  }
+
+  if (Object.keys(filterData).length > 0) {
+    andConditions.push({
+      AND: Object.entries(filterData).map(([field, value]) => ({
+        [field]:
+          value === 'true' ? true : value === 'false' ? false : Number(value),
+      })),
+    });
+  }
+
+  const whereConditions: Prisma.LocationWhereInput =
+    andConditions.length > 0 ? { AND: andConditions } : {};
+
   const result = await prisma.location.findMany({
+    where: whereConditions,
     orderBy: {
-      label: 'asc',
+      [sortBy]: sortOrder,
+    },
+    skip,
+    take: limit,
+    include: {
+      area: true,
     },
   });
 
-  return result;
+  const total = await prisma.location.count({
+    where: whereConditions,
+  });
+  const totalPage = Math.ceil(total / limit);
+
+  return {
+    meta: {
+      page,
+      limit,
+      total,
+      totalPage,
+    },
+    data: result,
+  };
 };
 
 // get single Location
@@ -31,6 +86,7 @@ const getSingleLocation = async (id: number): Promise<Location | null> => {
     where: {
       id,
     },
+    include: { area: true },
   });
 
   return result;
@@ -73,10 +129,20 @@ const deleteLocation = async (id: number): Promise<Location | null> => {
     where: {
       id,
     },
+    include: {
+      employees: true,
+    },
   });
 
   if (!isExist) {
     throw new ApiError(httpStatus.NOT_FOUND, 'Location Not Found');
+  }
+
+  if (isExist.employees.length) {
+    throw new ApiError(
+      httpStatus.BAD_REQUEST,
+      `Location Engaged with ${isExist.employees.length} Docs`
+    );
   }
 
   const result = await prisma.location.delete({
